ai_platform_engineering/agents/azuredevops/pyproject.toml
[project]
name = "agent_azuredevops"
version = "0.1.0"
description = "An Azure DevOps natural language agent using LangChain, LangGraph, and MCP."
authors = [
  {name = "CNOE Contributors", email = "cnoe@example.com"}
]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
  "a2a-sdk==0.3.0",
  "agntcy-app-sdk==0.4.0",
  "click==8.3.1",
  "cnoe-agent-utils==0.3.9",
  "langchain-core==1.1.2",
  "langchain-mcp-adapters==0.2.1",
  "langgraph==1.0.4",
  "mcp-azuredevops",
  "prometheus-client==0.23.1",
  "python-dotenv==1.2.1",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.metadata]
allow-direct-references = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.2",
]

[tool.ruff]
line-length = 140
indent-width = 2

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["F403"]

[tool.pytest.ini_options]
markers = [
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
]

[tool.uv.sources]
mcp-azuredevops = { path = "./mcp" }
